bar chart with add button
